ingestion-agent/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
python-multipart
